lib/rcfpch: change handling of missing RPC server
We have cases where RPC server is temporary not available, because RPC
provider is being restarted. While a monitor test or other user is
running and trying to access it.

In this case we don't want to log an error and polute the log,
moreover we want to return something sensible and distinguish
this case from other errors.

TE_ENOENT is returned in various cases while TE_ESRCH looks closer
to reality.

Signed-off-by: Konstantin Ushakov <kushakov@oktet.co.il>
Acked-by: Yurij Plotnikov <yurij.plotnikov@arknetworks.am>