trc: add ability to set unknown test/iteration handling
It's a long living issue that in a number of
suites default behaviour is PASSED will always be
treated as ok.

This patch adds an ability to parameterize this
behaviour on a per-TRC database basis. Keeping the
default behaviour as PASSED_UNKNOWN, i.e. treat
PASSED that are not in the DB as unknown.

Link: https://github.com/oktetlabs/test-environment/pull/66
Signed-off-by: Konstantin Ushakov <kushakov@oktet.co.il>
Acked-by: Yurij Plotnikov <yurij.plotnikov@arknetworks.am>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>