suites/selftest: fix PRNG selftest
It turns out that the Kolmogorov-Smirnov test that was
originally used to validate the uniformity of the PRNG
has some caveats when used on discrete distributions
that lead to an excessive number of false negatives.
The patch replaces it with a more universal Pearson's
chi-square test with its signifance level chosen in such
a way as to minimize the chance of false negative ands
still allow to catch genuine errors.

Fixes: bd996be73ceb ("lib/tools: provide a high-quality RNG")
OL-Redmine-Id: 14348
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>