lib/tapi_net: introduce lib to define and set up network
A Test Suite may require a flexible network setup
to support complex scenarios. The proposed API
provides a way to describe network topology using
a hierarchical structure of interfaces and the links
between them. All the required information is then
stored in the network context and used to set up the
specified network on test agents.

These changes introduce the initial functionality and
structure definitions required for this API.

Signed-off-by: Denis Pryazhennikov <denis.pryazhennikov@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>