lib/tapi: fix format args composing twice for verdicts
Usage of format string with arguments twice in the macro results
in execution of arguments composition twice. For example, if
function is used in arguments, it is executed twice. As the result
if the function is used to print something to buffer, it appears
twice in the verdict registered by Tester as MI log message.

OL-Redmine-Id: 14662
Signed-off-by: Andrew Rybchenko <andrew.rybchenko@oktetlabs.ru>
Reviewed-by: Artem Andreev <artem.andreev@oktetlabs.ru>